feat: use local index
