+++ templates/types/streaming/nextjs/app/api/chat/llamaindex/events.ts
@@ -3,26 +3,9 @@ import {
   CallbackManager,
   Metadata,
   NodeWithScore,
-  TextNode,
   ToolCall,
   ToolOutput,
 } from "llamaindex";
-
-function getNodeUrl(metadata: Metadata) {
-  const url = metadata["URL"];
-  if (url) return url;
-  const fileName = metadata["file_name"];
-  if (!process.env.FILESERVER_URL_PREFIX) {
-    console.warn(
-      "FILESERVER_URL_PREFIX is not set. File URLs will not be generated.",
-    );
-    return undefined;
-  }
-  if (fileName) {
-    return `${process.env.FILESERVER_URL_PREFIX}/data/${fileName}`;
-  }
-  return undefined;
-}
 
 export function appendSourceData(
   data: StreamData,
@@ -114,12 +97,18 @@ export function createCallbackManager(stream: StreamData) {
   return callbackManager;
 }
 
-export type DocumentFileType = "csv" | "pdf" | "txt" | "docx";
-
-export type DocumentFile = {
-  id: string;
-  filename: string;
-  filesize: number;
-  filetype: DocumentFileType;
-  content: string | TextNode[];
-};
+function getNodeUrl(metadata: Metadata) {
+  const url = metadata["URL"];
+  if (url) return url;
+  const fileName = metadata["file_name"];
+  if (!process.env.FILESERVER_URL_PREFIX) {
+    console.warn(
+      "FILESERVER_URL_PREFIX is not set. File URLs will not be generated.",
+    );
+    return undefined;
+  }
+  if (fileName) {
+    return `${process.env.FILESERVER_URL_PREFIX}/data/${fileName}`;
+  }
+  return undefined;
+}
