+++ templates/types/streaming/nextjs/app/api/chat/llamaindex/stream.ts
@@ -1,0 +1,40 @@
+import {
+  StreamData,
+  createCallbacksTransformer,
+  createStreamDataTransformer,
+  trimStartOfStreamHelper,
+  type AIStreamCallbacksAndOptions,
+} from "ai";
+import { EngineResponse } from "llamaindex";
+
+export function LlamaIndexStream(
+  response: AsyncIterable<EngineResponse>,
+  data: StreamData,
+  opts?: {
+    callbacks?: AIStreamCallbacksAndOptions;
+  },
+): ReadableStream<Uint8Array> {
+  return createParser(response, data)
+    .pipeThrough(createCallbacksTransformer(opts?.callbacks))
+    .pipeThrough(createStreamDataTransformer());
+}
+
+function createParser(res: AsyncIterable<EngineResponse>, data: StreamData) {
+  const it = res[Symbol.asyncIterator]();
+  const trimStartOfStream = trimStartOfStreamHelper();
+
+  return new ReadableStream<string>({
+    async pull(controller): Promise<void> {
+      const { value, done } = await it.next();
+      if (done) {
+        controller.close();
+        data.close();
+        return;
+      }
+      const text = trimStartOfStream(value.delta ?? "");
+      if (text) {
+        controller.enqueue(text);
+      }
+    },
+  });
+}
