+++ templates/types/streaming/nextjs/package.json
@@ -24,10 +24,9 @@
     "duck-duck-scrape": "^2.2.5",
     "formdata-node": "^6.0.3",
     "got": "^14.4.1",
-    "llamaindex": "0.4.5",
+    "llamaindex": "0.4.6",
     "lucide-react": "^0.294.0",
     "next": "^14.2.4",
-    "pdf2json": "3.0.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-markdown": "^8.0.7",
